refactor: move broadcast under utils
broadcast is a generic pub/sub ring buffer with zero juno dependencies
(stdlib only). It belongs under utils/ alongside other cross-cutting
utilities rather than at root level.
